Hinge Joint
-> 1 DOF hinge joint
-> Linear and angular constraints solved separately
-> Linear part is same as in position joint
-> Angular part is solved by projecting relative angular velocity on the
hige axis
-> Angular position correction is based on small angle approximation

Tested hinge joint in several configurations:
Door hinge, Ragdoll limbs, suspension bridge

Stable in most configurations which are setup carefully not to violate
the angle around an axis which is not the hinge axis

BridgeHinge tests a particularly unstable configuraion

The links are connected in series, such that the configuration is
hanging horizontally, and the weight of the links tries to rotate
connected bodies around an axis other than the hinge axis

Upto 5 links is very unstable with < 10 iterations
Somewhat stable with 100 iterations
Very stable with 500 iterations!!

Obviously something is very wrong. The culprit is ofcourse the angular
constraint. Small angle approximation i think.

Stability depends on how the joints are setup
More stable when hinge axis is closer to the bodies
More stable when link mass is distributed mostly around the hinge axis.